feat(chat): implement global chat UI with channels, messages, and user list
- Add ChatStore for centralized state management with events
- Create ChatOverlay with collapsible UI, drag support, and mobile tabs
- Implement ChannelList, MessageList, UserList, and ChatInput components
- Add message rendering for schematics, images, maps, and player connect links
- Integrate with ChatService for fetching messages, users, and sending messages
- Support replying to messages and displaying unread counts
