--- src/mindustrytool/features/chat/global/ChatOverlay.java
@@ -1,1475 +1,0 @@
-package mindustrytool.features.chat.global;
-
-import arc.Core;
-import arc.graphics.Color;
-import arc.graphics.g2d.Draw;
-import arc.graphics.g2d.Fill;
-import arc.scene.event.Touchable;
-import arc.scene.ui.Button;
-import arc.scene.ui.Image;
-import arc.scene.ui.Label;
-import arc.scene.ui.ScrollPane;
-import arc.scene.ui.TextButton;
-import arc.scene.ui.TextField;
-import arc.scene.ui.ImageButton.ImageButtonStyle;
-import arc.scene.ui.layout.Cell;
-import arc.scene.ui.layout.Scl;
-import arc.scene.ui.layout.Stack;
-import arc.scene.ui.layout.Table;
-import arc.struct.ObjectMap;
-import arc.struct.Seq;
-import mindustrytool.features.chat.global.dto.ChannelDto;
-import arc.util.Align;
-import arc.util.Log;
-import arc.util.Scaling;
-import arc.util.Time;
-import arc.util.Timer;
-import mindustry.Vars;
-import mindustry.game.EventType;
-import mindustry.game.Schematic;
-import mindustry.game.Schematics;
-import mindustry.game.EventType.ResizeEvent;
-import mindustry.game.EventType.Trigger;
-import mindustry.gen.Icon;
-import mindustry.gen.Tex;
-import mindustry.ui.Styles;
-import mindustry.ui.dialogs.SchematicsDialog.SchematicImage;
-import arc.Events;
-import arc.func.Prov;
-import mindustrytool.MdtKeybinds;
-import mindustrytool.Utils;
-import mindustrytool.features.auth.AuthService;
-import mindustrytool.features.auth.dto.LoginEvent;
-import mindustrytool.features.auth.dto.LogoutEvent;
-import mindustrytool.features.auth.dto.SessionLoadEvent;
-import mindustrytool.features.browser.map.MapDialog;
-import mindustrytool.features.browser.map.MapImage;
-import mindustrytool.features.browser.map.MapInfoDialog;
-import mindustrytool.features.browser.schematic.SchematicDialog;
-import mindustrytool.features.browser.schematic.SchematicInfoDialog;
-import mindustrytool.features.chat.global.dto.ChatMessage;
-import mindustrytool.features.chat.global.dto.ChatStateChange;
-import mindustrytool.features.chat.global.dto.ChatUser;
-import mindustrytool.services.UserService;
-import mindustrytool.ui.NetworkImage;
-import arc.scene.event.InputEvent;
-import arc.scene.event.InputListener;
-import arc.input.KeyCode;
-
-import java.time.Instant;
-import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
-import java.util.Arrays;
-import java.util.concurrent.CompletableFuture;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-import mindustrytool.features.chat.global.dto.ChatUser.SimpleRole;
-import mindustrytool.features.playerconnect.PlayerConnectLink;
-import mindustrytool.features.playerconnect.PlayerConnectRenderer;
-import mindustrytool.services.MapService;
-import mindustrytool.services.PlayerConnectService;
-import mindustrytool.services.SchematicService;
-import mindustrytool.services.State;
-
-public class ChatOverlay extends Table {
-    private static final Pattern MINDUSTRY_TOOL_LINK_PATTERN = Pattern
-            .compile("^https?://[^/]+/[^/]+/(schematics|maps)/([0-9a-fA-F-]+)");
-
-    private static final float TARGET_WIDTH = 250f;
-    private static final float PREVIEW_BUTTON_SIZE = 50f;
-    private static final float CARD_HEIGHT = 330f;
-    private static final String CURRENT_CHANNEL_ID_KEY = "mindustrytool-current-channel-id";
-
-    public enum MobileTab {
-        CHANNELS, MESSAGES, MEMBERS
-    }
-
-    private MobileTab currentMobileTab = MobileTab.MESSAGES;
-
-    private ObjectMap<String, Seq<ChatMessage>> messagesByChannel = new ObjectMap<>();
-    private arc.struct.ObjectSet<String> fullyLoadedChannels = new arc.struct.ObjectSet<>();
-    private ObjectMap<String, Integer> unreadByChannel = new ObjectMap<>();
-    private Seq<ChannelDto> channels = new Seq<>();
-    private String currentChannelId = Core.settings.getString(CURRENT_CHANNEL_ID_KEY, null);
-    private Table channelListTable;
-
-    private Table messageTable;
-    private Table userListTable;
-    private ScrollPane scrollPane;
-    private Table loadingTable;
-    private Table inputTable;
-    private TextField inputField;
-    private TextButton sendButton;
-
-    private State<Boolean> isSending = new State<>(false);
-    private boolean isLoadingMessages = false;
-
-    private String expandedMessageId = null;
-    private String replyingToMessageId = null;
-
-    private Table container;
-
-    private Cell<Table> containerCell;
-    private final PlayerConnectService playerConnectService = new PlayerConnectService();
-
-    private int unreadCount = 0;
-    private Table badgeTable;
-
-    private boolean isUserListCollapsed;
-    private Image connectionIndicator;
-
-    private final SchematicInfoDialog schematicInfoDialog = new SchematicInfoDialog();
-    private final MapInfoDialog mapInfoDialog = new MapInfoDialog();
-    private AttachContentDialog attachContentDialog;
-
-    public ChatOverlay() {
-        name = "mdt-chat-overlay";
-        touchable = Touchable.childrenOnly;
-        isUserListCollapsed = Vars.mobile;
-
-        setPosition(ChatConfig.x(), ChatConfig.y());
-
-        container = new Table();
-        inputTable = new Table();
-
-        containerCell = add(container);
-
-        visible(() -> Vars.ui.hudfrag != null && Vars.ui.hudfrag.shown);
-
-        inputField = new TextField();
-        inputField.setMessageText("@chat.enter-message");
-        inputField.setValidator(this::isValidInput);
-        inputField.keyDown(arc.input.KeyCode.enter, this::handleSend);
-
-        inputField.keyDown(arc.input.KeyCode.escape, this::collapse);
-
-        Events.on(EventType.ResizeEvent.class, e -> {
-            setup();
-            keepInScreen();
-        });
-
-        Events.on(LoginEvent.class, e -> Core.app.post(() -> setup()));
-        Events.on(LogoutEvent.class, e -> Core.app.post(() -> setup()));
-        Events.on(ResizeEvent.class, (event) -> {
-            setup();
-        });
-
-        Events.run(Trigger.update, () -> {
-            boolean noInputFocused = !Core.scene.hasField();
-
-            if (noInputFocused && Core.input.keyRelease(MdtKeybinds.chatKb)) {
-                ChatConfig.collapsed(!ChatConfig.collapsed());
-                setup();
-            }
-        });
-
-        addListener(new InputListener() {
-            @Override
-            public boolean keyDown(InputEvent event, KeyCode keycode) {
-                try {
-                    if (keycode == KeyCode.escape && !ChatConfig.collapsed()) {
-                        collapse();
-                        return true;
-                    }
-                } catch (Exception e) {
-                    Log.err(e);
-                }
-                return false;
-            }
-        });
-
-        Events.on(SessionLoadEvent.class, e -> {
-            if (!e.isLoading) {
-                buildInputTable(inputTable);
-            }
-        });
-
-        isSending.subscribe((curr, old) -> {
-            inputField.setDisabled(curr);
-            buildInputTable(inputTable);
-        });
-        Events.on(ChatStateChange.class, event -> {
-            var connected = event.connected;
-
-            if (connectionIndicator != null) {
-                connectionIndicator.setColor(connected ? Color.green : Color.yellow);
-            }
-        });
-        Core.app.post(() -> setup());
-    }
-
-    public void handleSend() {
-        boolean isSchematic = isSchematic(inputField.getText().trim());
-
-        if (isSchematic) {
-            sendSchematic();
-        } else if (inputField.isValid()) {
-            sendMessage();
-        }
-    }
-
-    public boolean isCollapsed() {
-        return ChatConfig.collapsed();
-    }
-
-    public void rebuild() {
-        setup();
-    }
-
-    private synchronized void setup() {
-        setPosition(ChatConfig.x(ChatConfig.collapsed()), ChatConfig.y(ChatConfig.collapsed()));
-
-        container.clearChildren();
-        container.touchable = Touchable.enabled;
-        container.setColor(1f, 1f, 1f, ChatConfig.opacity());
-
-        float scale = ChatConfig.scale();
-        float widthScale = ChatConfig.width();
-
-        if (ChatConfig.collapsed()) {
-            container.background(null);
-            containerCell.size(48);
-
-            Table buttonTable = new Table();
-            buttonTable.background(Styles.black6);
-
-            Button btn = new Button(Styles.clearNoneTogglei);
-            Stack stack = new Stack();
-            stack.add(new Image(Icon.chat));
-
-            badgeTable = new Table();
-            stack.add(badgeTable);
-            updateBadge();
-
-            btn.add(stack);
-            btn.addListener(new InputListener() {
-                float lastX, lastY;
-                boolean wasDragged = false;
-
-                @Override
-                public boolean touchDown(InputEvent event, float x, float y, int pointer, KeyCode button) {
-                    lastX = x;
-                    lastY = y;
-                    wasDragged = false;
-                    return true;
-                }
-
-                @Override
-                public void touchDragged(InputEvent event, float x, float y, int pointer) {
-                    try {
-                        float dx = x - lastX;
-                        float dy = y - lastY;
-
-                        if (Math.abs(dx) > 0.1f || Math.abs(dy) > 0.1f) {
-                            wasDragged = true;
-                        }
-
-                        ChatOverlay.this.moveBy(dx, dy);
-                        ChatConfig.x(ChatOverlay.this.x);
-                        ChatConfig.y(ChatOverlay.this.y);
-
-                        keepInScreen();
-                    } catch (Exception e) {
-                        Log.err(e);
-                    }
-                }
-
-                @Override
-                public void touchUp(InputEvent event, float x, float y, int pointer, KeyCode button) {
-                    if (wasDragged) {
-                        return;
-                    }
-
-                    try {
-                        ChatConfig.collapsed(false);
-                        unreadCount = 0;
-                        Core.app.post(() -> setup());
-                    } catch (Exception e) {
-                        Log.err(e);
-                    }
-                }
-            });
-
-            buttonTable.add(btn).grow();
-            container.add(buttonTable).grow();
-        } else {
-            container.background(Styles.black8);
-            float width = Core.graphics.getWidth() / Scl.scl() * 0.9f * widthScale;
-            float height = Core.graphics.getHeight() / Scl.scl() * 0.9f * scale;
-
-            float actualWidth = Math.min(width, 1900f * widthScale);
-            containerCell.size(actualWidth, Math.min(height, 1300f * scale));
-
-            // Header
-            Table header = new Table();
-            header.background(Styles.black6);
-            header.touchable(() -> Touchable.enabled);
-
-            // Header Drag Listener
-            header.addListener(new InputListener() {
-                float lastX, lastY;
-
-                @Override
-                public boolean touchDown(InputEvent event, float x, float y, int pointer, KeyCode button) {
-                    lastX = x;
-                    lastY = y;
-                    return true;
-                }
-
-                @Override
-                public void touchDragged(InputEvent event, float x, float y, int pointer) {
-                    try {
-                        ChatOverlay.this.moveBy(x - lastX, y - lastY);
-                        keepInScreen();
-                        ChatConfig.x(ChatOverlay.this.x);
-                        ChatConfig.y(ChatOverlay.this.y);
-                    } catch (Exception e) {
-                        Log.info(e);
-                    }
-                }
-            });
-
-            connectionIndicator = new Image(Tex.whiteui) {
-                @Override
-                public void draw() {
-                    Draw.color(color);
-                    Fill.circle(x + width / 2f, y + height / 2f, Math.min(width, height) / 2f);
-                    Draw.reset();
-                }
-            };
-            connectionIndicator.setColor(ChatService.getInstance().isConnected() ? Color.green : Color.yellow);
-
-            if (Vars.mobile) {
-                if (currentMobileTab == MobileTab.MESSAGES) {
-                    header.button(Icon.left, Styles.clearNonei, () -> {
-                        currentMobileTab = MobileTab.CHANNELS;
-                        setup();
-                    }).size(40 * scale).padLeft(4 * scale);
-                    header.image(Icon.move).color(Color.gray).size(24 * scale).padLeft(4 * scale);
-
-                    String chanName = "@chat.global-chat";
-                    if (currentChannelId != null) {
-                        ChannelDto c = channels.find(ch -> ch.id.equals(currentChannelId));
-                        if (c != null)
-                            chanName = "# " + c.name;
-                    }
-                    Label title = header.add(chanName).style(Styles.outlineLabel).padLeft(8 * scale).get();
-                    title.setFontScale(scale);
-
-                    header.add(connectionIndicator).size(10 * scale).padLeft(8 * scale);
-                    header.add().growX();
-
-                    header.button(Icon.right, Styles.clearNonei, () -> {
-                        currentMobileTab = MobileTab.MEMBERS;
-                        setup();
-                    }).size(40 * scale).padRight(4 * scale);
-                    header.button(Icon.refresh, Styles.clearNonei, () -> {
-                        messagesByChannel.clear();
-                        fullyLoadedChannels.clear();
-                    }).size(40 * scale).padRight(4 * scale);
-                    header.button(Icon.cancel, Styles.clearNonei, this::collapse).size(40 * scale).padRight(4 * scale);
-                } else if (currentMobileTab == MobileTab.CHANNELS) {
-                    header.image(Icon.move).color(Color.gray).size(24 * scale).padLeft(8 * scale);
-                    Label title = header.add("@chat.channels").style(Styles.outlineLabel).padLeft(8 * scale).get();
-                    title.setFontScale(scale);
-                    header.add().growX();
-                    header.button(Icon.refresh, Styles.clearNonei, () -> {
-                        messagesByChannel.clear();
-                        fullyLoadedChannels.clear();
-                    }).size(40 * scale).padRight(4 * scale);
-                    header.button(Icon.cancel, Styles.clearNonei, this::collapse).size(40 * scale).padRight(4 * scale);
-                    header.button(Icon.right, Styles.clearNonei, () -> {
-                        currentMobileTab = MobileTab.MESSAGES;
-                        setup();
-                    }).size(40 * scale).padRight(4 * scale);
-                } else if (currentMobileTab == MobileTab.MEMBERS) {
-                    header.button(Icon.left, Styles.clearNonei, () -> {
-                        currentMobileTab = MobileTab.MESSAGES;
-                        setup();
-                    }).size(40 * scale).padLeft(4 * scale);
-                    header.image(Icon.move).color(Color.gray).size(24 * scale).padLeft(4 * scale);
-                    Label title = header.add("@chat.online-members").style(Styles.outlineLabel).padLeft(8 * scale)
-                            .get();
-                    title.setFontScale(scale);
-                    header.add().growX();
-                    header.button(Icon.refresh, Styles.clearNonei, () -> {
-                        messagesByChannel.clear();
-                        fullyLoadedChannels.clear();
-                    }).size(40 * scale).padRight(4 * scale);
-                    header.button(Icon.cancel, Styles.clearNonei, this::collapse).size(40 * scale).padRight(4 * scale);
-                }
-            } else {
-                header.image(Icon.move).color(Color.gray).size(24 * scale).padLeft(8 * scale);
-                Label title = header.add("@chat.global-chat").style(Styles.outlineLabel).padLeft(8 * scale).get();
-                title.setFontScale(scale);
-
-                header.add(connectionIndicator).size(10 * scale).padLeft(8 * scale);
-
-                header.add().growX();
-
-                // Minimize button
-                header.button(Icon.refresh, Styles.clearNonei, () -> {
-                    messagesByChannel.clear();
-                    fullyLoadedChannels.clear();
-                }).size(40 * scale).padRight(4 * scale);
-                header.button(Icon.cancel, Styles.clearNonei, this::collapse).size(40 * scale).padRight(4 * scale);
-            }
-
-            container.add(header).growX().height(46 * scale).row();
-
-            // Main Content Area
-            Table mainContent = new Table();
-
-            if (Vars.mobile) {
-                if (currentMobileTab == MobileTab.CHANNELS) {
-                    Table leftSide = new Table();
-                    leftSide.top().background(Styles.black5);
-                    channelListTable = new Table();
-                    channelListTable.top().left();
-                    ScrollPane channelScroll = new ScrollPane(channelListTable, Styles.noBarPane);
-                    channelScroll.setScrollingDisabled(true, false);
-                    leftSide.add(channelScroll).grow();
-                    mainContent.add(leftSide).grow();
-                } else if (currentMobileTab == MobileTab.MEMBERS) {
-                    Table rightSide = new Table();
-                    rightSide.top().background(Styles.black3);
-                    userListTable = new Table();
-                    userListTable.top().left();
-                    ScrollPane userScrollPane = new ScrollPane(userListTable, Styles.noBarPane);
-                    userScrollPane.setScrollingDisabled(true, false);
-                    rightSide.add(userScrollPane).grow().row();
-                    mainContent.add(rightSide).grow();
-                } else {
-                    messageTable = new Table();
-                    messageTable.top().left();
-
-                    scrollPane = new ScrollPane(messageTable, Styles.noBarPane);
-                    scrollPane.setScrollingDisabled(true, false);
-                    scrollPane.setFadeScrollBars(false);
-                    scrollPane.visible(() -> ChatService.getInstance().isConnected());
-                    scrollPane.scrolled(amount -> {
-                        if (scrollPane.getScrollY() < 100 && !isLoadingMessages && currentChannelId != null
-                                && !fullyLoadedChannels.contains(currentChannelId)) {
-                            Seq<ChatMessage> msgs = messagesByChannel.get(currentChannelId);
-                            if (msgs != null && !msgs.isEmpty()) {
-                                loadMoreMessages(currentChannelId, msgs.first().id);
-                            }
-                        }
-                    });
-
-                    loadingTable = new Table();
-
-                    loadingTable.add("@loading").style(Styles.defaultLabel).color(Color.gray)
-                            .visible(() -> !ChatService.getInstance().isConnected() || isLoadingMessages).get()
-                            .setFontScale(scale);
-
-                    Stack stack = new Stack();
-                    stack.add(loadingTable);
-                    stack.add(scrollPane);
-
-                    Table centerArea = new Table();
-                    centerArea.add(stack).grow().row();
-
-                    buildInputTable(inputTable);
-                    centerArea.add(inputTable).growX().bottom();
-
-                    mainContent.add(centerArea).grow().minWidth(0);
-                }
-                container.add(mainContent).grow().minWidth(0).row();
-            } else {
-                float leftWidth = Math.min(160f * scale, actualWidth * 0.25f);
-                float rightWidthExp = Math.min(280f * scale, actualWidth * 0.35f);
-                float rightWidthCol = 48f * scale;
-
-                // Left Side - Channels
-                Table leftSide = new Table();
-                leftSide.top().background(Styles.black5);
-                channelListTable = new Table();
-                channelListTable.top().left();
-                ScrollPane channelScroll = new ScrollPane(channelListTable, Styles.noBarPane);
-                channelScroll.setScrollingDisabled(true, false);
-                leftSide.add(channelScroll).grow();
-
-                mainContent.add(leftSide).width(leftWidth).growY();
-                mainContent.image(Tex.whiteui).width(1f * scale).color(Color.darkGray).fillY();
-
-                // Messages
-                messageTable = new Table();
-                messageTable.top().left();
-
-                scrollPane = new ScrollPane(messageTable, Styles.noBarPane);
-                scrollPane.setScrollingDisabled(true, false);
-                scrollPane.setFadeScrollBars(false);
-                scrollPane.visible(() -> ChatService.getInstance().isConnected());
-                scrollPane.scrolled(amount -> {
-                    if (scrollPane.getScrollY() < 100 && !isLoadingMessages && currentChannelId != null
-                            && !fullyLoadedChannels.contains(currentChannelId)) {
-                        Seq<ChatMessage> msgs = messagesByChannel.get(currentChannelId);
-                        if (msgs != null && !msgs.isEmpty()) {
-                            loadMoreMessages(currentChannelId, msgs.first().id);
-                        }
-                    }
-                });
-
-                loadingTable = new Table();
-
-                loadingTable.add("@loading").style(Styles.defaultLabel).color(Color.gray)
-                        .visible(() -> !ChatService.getInstance().isConnected() || isLoadingMessages).get()
-                        .setFontScale(scale);
-
-                Stack stack = new Stack();
-                stack.add(loadingTable);
-                stack.add(scrollPane);
-
-                Table centerArea = new Table();
-                centerArea.add(stack).grow().row();
-
-                buildInputTable(inputTable);
-                centerArea.add(inputTable).growX().bottom();
-
-                mainContent.add(centerArea).grow().minWidth(0);
-
-                // Vertical Separator
-                mainContent.image(Tex.whiteui).width(1f * scale).color(Color.darkGray).fillY();
-
-                // User List Sidebar
-                Table rightSide = new Table();
-                rightSide.top();
-                rightSide.background(Styles.black3);
-
-                userListTable = new Table();
-                userListTable.top().left();
-                ScrollPane userScrollPane = new ScrollPane(userListTable, Styles.noBarPane);
-                userScrollPane.setScrollingDisabled(true, false);
-
-                Runnable rebuildRightSide = new Runnable() {
-                    @Override
-                    public void run() {
-                        rightSide.clear();
-                        Table titleTable = new Table();
-                        if (!isUserListCollapsed) {
-                            Label l = titleTable.add("@chat.online-members").style(Styles.defaultLabel)
-                                    .color(Color.gray)
-                                    .pad(10 * scale).left()
-                                    .minWidth(0).ellipsis(true).growX().get();
-                            l.setFontScale(scale);
-                        }
-
-                        titleTable.button(isUserListCollapsed ? Icon.left : Icon.right, Styles.clearNonei, () -> {
-                            isUserListCollapsed = !isUserListCollapsed;
-                            this.run();
-                            Cell<?> cell = mainContent.getCell(rightSide);
-                            if (cell != null) {
-                                cell.width(isUserListCollapsed ? rightWidthCol : rightWidthExp);
-                            }
-                        }).size(40 * scale).pad(4 * scale).right();
-
-                        rightSide.add(titleTable).growX().row();
-
-                        if (!isUserListCollapsed) {
-                            rightSide.add(userScrollPane).grow().row();
-                        }
-                    }
-                };
-                rebuildRightSide.run();
-
-                mainContent.add(rightSide).width(isUserListCollapsed ? rightWidthCol : rightWidthExp).growY();
-
-                container.add(mainContent).grow().minWidth(0).row();
-            }
-
-            // Fetch users
-            Timer.schedule(() -> {
-                if (currentChannelId != null) {
-                    ChatService.getInstance().getChatUsers(currentChannelId, this::rebuildUserList,
-                            e -> Log.err("Failed to fetch users", e));
-                }
-            }, 0.25f);
-
-            if (channels.isEmpty()) {
-                ChatService.getInstance().getChannels(chans -> {
-                    channels.clear();
-                    channels.addAll(chans);
-                    if (channels.size > 0) {
-                        if (currentChannelId == null || !channels.contains(c -> c.id.equals(currentChannelId))) {
-                            switchChannel(channels.get(0).id);
-                        } else {
-                            switchChannel(currentChannelId);
-                        }
-                    }
-                    rebuildChannelList();
-                }, e -> Log.err("Failed to fetch channels", e));
-            } else {
-                rebuildChannelList();
-            }
-
-            // Initial population
-            rebuildMessages(messageTable);
-
-            // Scroll to bottom after layout
-            Core.app.post(() -> {
-                if (scrollPane != null) {
-                    scrollPane.setScrollY(scrollPane.getMaxY());
-                }
-            });
-
-            Time.runTask(60, () -> {
-                if (scrollPane != null) {
-                    Core.app.post(() -> scrollPane.setScrollY(scrollPane.getMaxY()));
-                }
-            });
-
-        }
-
-        pack();
-
-        keepInScreen();
-
-    }
-
-    private void rebuildChannelList() {
-        if (channelListTable == null)
-            return;
-        channelListTable.clear();
-        channelListTable.top().left();
-
-        float scale = ChatConfig.scale();
-
-        for (ChannelDto channel : channels) {
-            boolean isSelected = channel.id.equals(currentChannelId);
-            TextButton btn = new TextButton("# " + channel.name, isSelected ? Styles.togglet : Styles.cleart);
-            btn.getLabel().setAlignment(Align.left);
-            btn.getLabel().setFontScale(scale);
-            btn.getLabel().setEllipsis(true);
-            if (btn.getLabelCell() != null)
-                btn.getLabelCell().minWidth(0);
-
-            if (isSelected) {
-                btn.setChecked(true);
-            }
-
-            int unread = unreadByChannel.get(channel.id, 0);
-            if (unread > 0 && !isSelected) {
-                btn.getLabel().setColor(Color.white);
-                btn.setText("# " + channel.name + " (" + (unread > 99 ? "99+" : unread) + ")");
-            } else {
-                btn.getLabel().setColor(Color.lightGray);
-            }
-
-            btn.clicked(() -> {
-                if (!isSelected) {
-                    switchChannel(channel.id);
-                }
-                if (Vars.mobile && currentMobileTab != MobileTab.MESSAGES) {
-                    currentMobileTab = MobileTab.MESSAGES;
-                    setup();
-                }
-            });
-
-            channelListTable.add(btn).growX().minWidth(0).height(40 * scale).pad(2 * scale).row();
-        }
-    }
-
-    private void switchChannel(String channelId) {
-        currentChannelId = channelId;
-        Core.settings.put(CURRENT_CHANNEL_ID_KEY, channelId);
-        unreadByChannel.put(channelId, 0);
-        updateBadge();
-        rebuildChannelList();
-
-        if (messageTable != null) {
-            messageTable.clear();
-        }
-
-        if (!messagesByChannel.containsKey(channelId) || messagesByChannel.get(channelId).isEmpty()) {
-            isLoadingMessages = true;
-            ChatService.getInstance().fetchMessages(channelId, null, msgs -> {
-                if (channelId.equals(currentChannelId)) {
-                    isLoadingMessages = false;
-                }
-                Seq<ChatMessage> seq = new Seq<>(msgs);
-                messagesByChannel.put(channelId, seq.reverse());
-                if (channelId.equals(currentChannelId)) {
-                    rebuildMessages(messageTable);
-                    if (scrollPane != null) {
-                        Core.app.post(() -> scrollPane.setScrollY(scrollPane.getMaxY()));
-                    }
-                }
-            }, e -> {
-                if (channelId.equals(currentChannelId)) {
-                    isLoadingMessages = false;
-                }
-                Log.err("Failed to fetch messages for channel", e);
-            });
-        } else {
-            isLoadingMessages = false;
-            rebuildMessages(messageTable);
-            if (scrollPane != null) {
-                Core.app.post(() -> scrollPane.setScrollY(scrollPane.getMaxY()));
-            }
-        }
-
-        ChatService.getInstance().getChatUsers(channelId, this::rebuildUserList,
-                e -> Log.err("Failed to fetch users", e));
-    }
-
-    private void loadMoreMessages(String channelId, String cursor) {
-        if (isLoadingMessages)
-            return;
-        isLoadingMessages = true;
-        ChatService.getInstance().fetchMessages(channelId, cursor, msgs -> {
-            if (channelId.equals(currentChannelId)) {
-                isLoadingMessages = false;
-            }
-            if (msgs.length > 0) {
-                Seq<ChatMessage> seq = messagesByChannel.get(channelId);
-                if (seq == null) {
-                    seq = new Seq<>();
-                    messagesByChannel.put(channelId, seq);
-                }
-
-                // Keep track of old scroll Y
-                float oldMaxY = scrollPane != null ? scrollPane.getMaxY() : 0;
-                float oldScrollY = scrollPane != null ? scrollPane.getScrollY() : 0;
-
-                Seq<ChatMessage> oldMsgs = new Seq<>(msgs);
-                oldMsgs.reverse().addAll(seq);
-                messagesByChannel.put(channelId, oldMsgs);
-
-                if (channelId.equals(currentChannelId)) {
-                    rebuildMessages(messageTable);
-                    if (scrollPane != null) {
-                        Core.app.post(() -> {
-                            float newMaxY = scrollPane.getMaxY();
-                            scrollPane.setScrollY(oldScrollY + (newMaxY - oldMaxY));
-                        });
-                    }
-                }
-            } else {
-                fullyLoadedChannels.add(channelId);
-            }
-        }, e -> {
-            if (channelId.equals(currentChannelId)) {
-                isLoadingMessages = false;
-            }
-            Log.err("Failed to fetch more messages for channel", e);
-        });
-    }
-
-    private void buildInputTable(Table inputTable) {
-        inputTable.clear();
-        inputTable.background(Styles.black6);
-
-        float scale = ChatConfig.scale();
-
-        if (AuthService.getInstance().isLoggedIn()) {
-            if (replyingToMessageId != null && currentChannelId != null) {
-                Seq<ChatMessage> msgs = messagesByChannel.get(currentChannelId);
-                ChatMessage repliedMsg = msgs != null ? msgs.find(m -> m.id.equals(replyingToMessageId)) : null;
-                if (repliedMsg != null) {
-                    Table replyContainer = new Table();
-                    replyContainer.background(Styles.black5);
-                    replyContainer.left();
-                    replyContainer.add(new Image(Icon.upSmall)).size(16 * scale).padRight(4 * scale).color(Color.gray);
-
-                    UserService.findUserById(repliedMsg.createdBy).thenAccept(replyData -> {
-                        Core.app.post(() -> {
-                            Label replyUser = new Label(replyData.getName());
-                            replyUser.setFontScale(scale * 0.8f);
-                            replyUser.setColor(Color.gray);
-                            replyContainer.add(replyUser).padRight(4 * scale);
-
-                            Label replyContent = new Label(repliedMsg.content.replace('\n', ' '));
-                            replyContent.setFontScale(scale * 0.8f);
-                            replyContent.setColor(Color.gray);
-                            replyContent.setEllipsis(true);
-                            replyContainer.add(replyContent).minWidth(0).maxWidth(200 * scale).padRight(8 * scale);
-
-                            replyContainer.button(Icon.cancel, Styles.clearNonei, () -> {
-                                replyingToMessageId = null;
-                                buildInputTable(inputTable);
-                            }).size(20 * scale);
-                        });
-                    });
-                    inputTable.add(replyContainer).growX().pad(4 * scale).row();
-                }
-            }
-
-            Table inputRow = new Table();
-            sendButton = new TextButton(isSending.get() ? "@sending" : "@chat.send", Styles.defaultt);
-            sendButton.clicked(this::handleSend);
-            sendButton.setDisabled(() -> isSending.get());
-
-            inputRow.add(inputField).growX().minWidth(0).height(40f * scale).pad(8 * scale).padRight(4 * scale);
-
-            inputRow.button(Utils.icons("attach-file.png"), () -> {
-                if (attachContentDialog == null) {
-                    attachContentDialog = new AttachContentDialog(this::handleAttachContent);
-                }
-                attachContentDialog.show();
-            }).pad(8 * scale).size(40f * scale);
-
-            inputRow.add(sendButton).width(100f * scale).height(40f * scale).pad(8 * scale).padLeft(0);
-            inputTable.add(inputRow).growX().minWidth(0);
-        } else {
-            inputTable.button("@login", Styles.defaultt, () -> {
-                AuthService.getInstance().login();
-            }).growX().height(40f * scale).pad(8 * scale);
-        }
-    }
-
-    private boolean isSchematic(String text) {
-        if (!text.startsWith(Vars.schematicBaseStart)) {
-            return false;
-        }
-
-        try {
-            Schematics.readBase64(text);
-            return true;
-        } catch (Exception _e) {
-            return false;
-        }
-    }
-
-    private boolean isValidInput(String text) {
-        if (text.length() <= 0) {
-            return false;
-        }
-
-        if (text.startsWith(Vars.schematicBaseStart)) {
-            if (text.length() > 2056 * 12) {
-                return false;
-            }
-        } else {
-            if (text.length() > 2056) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    public void keepInScreen() {
-        if (getScene() == null) {
-            return;
-        }
-
-        float w = getWidth();
-        float h = getHeight();
-        float sw = getScene().getWidth();
-        float sh = getScene().getHeight();
-
-        if (x < 0) {
-            x = 0;
-        }
-
-        if (y < 0) {
-            y = 0;
-        }
-
-        if (x + w > sw) {
-            x = sw - w;
-        }
-
-        if (y + h > sh) {
-            y = sh - h;
-        }
-    }
-
-    public synchronized void addMessages(ChatMessage... newMessages) {
-        int totalAddedCount = 0;
-        boolean currentChannelUpdated = false;
-
-        for (ChatMessage msg : newMessages) {
-            String cId = msg.channelId;
-            if (cId == null)
-                continue;
-
-            if (!messagesByChannel.containsKey(cId)) {
-                messagesByChannel.put(cId, new Seq<>());
-            }
-            Seq<ChatMessage> channelMsgs = messagesByChannel.get(cId);
-
-            if (channelMsgs.contains(m -> m.id.equals(msg.id))) {
-                continue;
-            }
-
-            channelMsgs.add(msg);
-
-            if (cId.equals(currentChannelId)) {
-                currentChannelUpdated = true;
-            }
-
-            if (channelMsgs.size > 1000) {
-                channelMsgs.remove(0);
-            }
-
-            try {
-                if ((ChatConfig.collapsed() || !cId.equals(currentChannelId))
-                        && ChatConfig.lastRead().isBefore(Instant.parse(msg.createdAt))) {
-                    unreadByChannel.put(cId, unreadByChannel.get(cId, 0) + 1);
-                    totalAddedCount++;
-                }
-            } catch (Exception e) {
-                Log.err(e);
-            }
-        }
-
-        if (totalAddedCount > 0) {
-            unreadCount += totalAddedCount;
-            updateBadge();
-            Core.app.post(() -> rebuildChannelList());
-        }
-
-        if (!ChatConfig.collapsed() && currentChannelId != null && currentChannelUpdated) {
-            ChatConfig.lastRead(Instant.now());
-        }
-
-        if (messageTable != null && !ChatConfig.collapsed() && currentChannelUpdated) {
-            Core.app.post(() -> {
-                rebuildMessages(messageTable);
-            });
-
-            Time.runTask(10, () -> {
-                if (scrollPane != null) {
-                    Core.app.post(() -> scrollPane.setScrollY(scrollPane.getMaxY()));
-                }
-            });
-        }
-    }
-
-    private void rebuildMessages(Table messageTable) {
-        if (messageTable == null)
-            return;
-        messageTable.clear();
-        messageTable.top().left();
-
-        float scale = ChatConfig.scale();
-
-        if (currentChannelId != null) {
-            ChatConfig.lastRead(Instant.now());
-        }
-
-        if (currentChannelId == null && !channels.isEmpty()) {
-            currentChannelId = channels.get(0).id;
-        }
-
-        if (currentChannelId == null) {
-            return;
-        }
-
-        Seq<ChatMessage> channelMsgs = messagesByChannel.get(currentChannelId);
-        if (channelMsgs == null)
-            return;
-
-        for (ChatMessage msg : channelMsgs) {
-            Table entry = new Table();
-
-            entry.setBackground(null);
-            entry.table(avatar -> {
-                avatar.top();
-                UserService.findUserById(msg.createdBy).thenAccept(data -> {
-                    Core.app.post(() -> {
-                        avatar.clear();
-                        if (data.getImageUrl() != null && !data.getImageUrl().isEmpty()) {
-                            avatar.add(new NetworkImage(data.getImageUrl())).size(40 * scale);
-                        } else {
-                            avatar.add(new Image(Icon.players)).size(40 * scale);
-                        }
-                    });
-                });
-            }).size(48 * scale).top().pad(8 * scale);
-
-            entry.table(card -> {
-                card.top().left();
-
-                Label label = new Label("...");
-                label.setStyle(Styles.defaultLabel);
-                label.setFontScale(scale);
-
-                UserService.findUserById(msg.createdBy).thenAccept(data -> {
-                    Core.app.post(() -> {
-                        String timeStr = msg.createdAt;
-
-                        if (msg.createdAt != null) {
-                            try {
-                                Instant instant = Instant.parse(msg.createdAt);
-                                timeStr = DateTimeFormatter.ofPattern("HH:mm")
-                                        .withZone(ZoneId.systemDefault())
-                                        .format(instant);
-
-                            } catch (Throwable err) {
-                                Log.err(err);
-                            }
-                        }
-
-                        Color color = data.getHighestRole()
-                                .map(r -> {
-                                    try {
-                                        return Color.valueOf(r.getColor());
-                                    } catch (Exception err) {
-                                        Log.err(err);
-                                        return Color.white;
-                                    }
-                                })
-                                .orElse(Color.white);
-
-                        label.setText("[#" + color.toString() + "]" + data.getName() + "[white]"
-                                + (timeStr.isEmpty() ? "" : " [gray]" + timeStr));
-                    });
-                });
-
-                card.add(label).left().row();
-                if (msg.replyTo != null && !msg.replyTo.isEmpty()) {
-                    ChatMessage repliedMsg = channelMsgs.find(m -> m.id.equals(msg.replyTo));
-                    if (repliedMsg != null) {
-                        card.table(replyTable -> {
-                            replyTable.center().left();
-                            replyTable.image(Icon.rightSmall).size(16 * scale).padRight(4 * scale)
-                                    .color(Color.gray);
-
-                            Label replyContent = new Label(repliedMsg.content.replace('\n', ' '));
-                            replyContent.setFontScale(scale);
-                            replyContent.setColor(Color.gray);
-                            replyContent.setEllipsis(true);
-                            replyTable.add(replyContent).minWidth(0).maxWidth(200 * scale);
-                        }).growX().padBottom(2 * scale).row();
-                    }
-                }
-                card.table(c -> {
-                    String content = msg.content.trim();
-
-                    if (PlayerConnectLink.isValid(content)) {
-                        Label l = c.add(content).wrap().color(Color.lightGray).left().growX().get();
-                        l.setFontScale(scale);
-                        c.row();
-                        renderPlayerConnectRoom(c, content);
-                        return;
-                    }
-
-                    if (NetworkImage.isValidImageLink(content)) {
-                        Label l = c.add(content).wrap().color(Color.lightGray).left().growX().get();
-                        l.setFontScale(scale);
-                        c.row();
-                        c.add(new NetworkImage(content)).maxHeight(800 * scale).maxWidth(800 * scale);
-                        c.table().growX();
-                        return;
-                    }
-
-                    int schematicBasePosition = content.indexOf(Vars.schematicBaseStart);
-
-                    if (schematicBasePosition != -1) {
-                        int endPosition = content.indexOf(" ", schematicBasePosition) + 1;
-
-                        if (endPosition == 0) {
-                            endPosition = content.length();
-                        }
-
-                        String prev = content.substring(0, schematicBasePosition);
-
-                        Label l = c.add(prev).wrap().color(Color.lightGray).left().growX().get();
-                        l.setFontScale(scale);
-                        String schematicBase64 = content.substring(schematicBasePosition, endPosition);
-
-                        try {
-                            var schematic = Schematics.readBase64(schematicBase64);
-                            c.row();
-                            renderSchematic(card, schematic);
-                            c.row();
-                            String after = content.substring(endPosition);
-                            Label l2 = c.add(after).wrap().color(Color.lightGray).left().growX().get();
-                            l2.setFontScale(scale);
-                        } catch (Exception e) {
-                            c.clear();
-                            Label l2 = c.add(content).wrap().color(Color.lightGray).left().growX().get();
-                            l2.setFontScale(scale);
-                        }
-
-                        return;
-                    }
-
-                    Matcher matcher = MINDUSTRY_TOOL_LINK_PATTERN.matcher(content);
-
-                    if (matcher.find()) {
-                        String url = matcher.group(0);
-                        String contentType = matcher.group(1);
-                        String id = matcher.group(2);
-
-                        if (contentType.equals("maps")) {
-                            renderMap(c, id, url);
-                            c.table().growX();
-                            return;
-                        }
-
-                        if (contentType.equals("schematics")) {
-                            renderSchematic(c, id, url);
-                            c.table().growX();
-                            return;
-                        }
-                    }
-
-                    Label l = c.add(content).wrap().color(Color.lightGray).left().growX().get();
-                    l.setFontScale(scale);
-                })
-                        .top().left().growX().padTop(6 * scale);
-
-                card.clicked(() -> {
-                    if (expandedMessageId != null && expandedMessageId.equals(msg.id)) {
-                        expandedMessageId = null;
-                    } else {
-                        expandedMessageId = msg.id;
-                    }
-                    rebuildMessages(messageTable);
-                });
-
-                if (expandedMessageId != null && expandedMessageId.equals(msg.id)) {
-                    card.row();
-                    card.table(actions -> {
-                        actions.left().defaults().height(36 * scale).minWidth(160 * scale).padRight(8 * scale);
-
-                        TextButton copyBtn = new TextButton("@copy", Styles.defaultt);
-                        copyBtn.clicked(() -> {
-                            try {
-                                Core.app.setClipboardText(msg.content);
-                                Vars.ui.showInfoFade("@copied");
-                                expandedMessageId = null;
-                                rebuildMessages(messageTable);
-                            } catch (Exception e) {
-                                Vars.ui.showInfoFade(e.getMessage());
-                            }
-                        });
-                        copyBtn.getLabel().setFontScale(scale * 0.8f);
-                        actions.add(copyBtn);
-
-                        TextButton replyBtn = new TextButton("@chat.reply", Styles.defaultt);
-                        replyBtn.clicked(() -> {
-                            replyingToMessageId = msg.id;
-                            expandedMessageId = null;
-                            rebuildMessages(messageTable);
-                            buildInputTable(inputTable);
-                        });
-                        replyBtn.getLabel().setFontScale(scale * 0.8f);
-                        actions.add(replyBtn);
-
-                    }).growX().padTop(4 * scale);
-                }
-            }).growX().pad(8 * scale).top();
-
-            messageTable.add(entry).growX().padBottom(4 * scale).row();
-        }
-    }
-
-    private void renderSchematic(Table table, String id, String url) {
-        table.table(Tex.pane, preview -> {
-            preview.top().left().margin(0f);
-
-            preview.table(buttons -> {
-                buttons.center().defaults().size(PREVIEW_BUTTON_SIZE);
-
-                buttons.button(Icon.copy, Styles.emptyi, () -> SchematicDialog.handleCopySchematic(id))
-                        .pad(2);
-
-                buttons.button(Icon.download, Styles.emptyi, () -> SchematicDialog
-                        .handleDownloadSchematic(id))
-                        .pad(2);
-
-                buttons.button(Icon.info, Styles.emptyi,
-                        () -> SchematicService.findSchematicById(id)
-                                .thenAccept(schem -> Core.app.post(() -> schematicInfoDialog.show(schem))))
-                        .tooltip("@info.title");
-
-                buttons.button(Icon.link, Styles.emptyi, () -> Core.app.openURI(url))
-                        .pad(2);
-            }).growX().height(PREVIEW_BUTTON_SIZE);
-
-            preview.row();
-
-            preview.stack(new Table(t -> t.add(new mindustrytool.features.browser.schematic.SchematicImage(id))))
-                    .top()
-                    .left();
-        }).style(Styles.flati).width(TARGET_WIDTH).height(CARD_HEIGHT).top()
-                .left();
-
-    }
-
-    private void renderMap(Table table, String id, String url) {
-        table.table(Tex.pane, preview -> {
-            preview.top().left().margin(0f);
-
-            preview.table(buttons -> {
-                buttons.center().defaults().size(PREVIEW_BUTTON_SIZE);
-
-                buttons.button(Icon.download, Styles.emptyi, () -> MapDialog.handleDownloadMap(id))
-                        .pad(2);
-
-                buttons.button(Icon.info, Styles.emptyi,
-                        () -> MapService.findMapById(id)
-                                .thenAccept(m -> Core.app.post(() -> mapInfoDialog.show(m))))
-                        .tooltip("@info.title");
-
-                buttons.button(Icon.link, Styles.emptyi, () -> Core.app.openURI(url))
-                        .pad(2);
-            }).growX().height(PREVIEW_BUTTON_SIZE);
-
-            preview.row();
-
-            preview.stack(new Table(t -> t.add(new MapImage(id)))).top().left();
-        }).style(Styles.flati).width(TARGET_WIDTH).height(CARD_HEIGHT)
-                .top()
-                .left();
-    }
-
-    private void rebuildUserList(ChatUser[] users) {
-        if (userListTable == null) {
-            return;
-        }
-
-        userListTable.clear();
-        userListTable.top().left();
-
-        float scale = ChatConfig.scale();
-
-        Arrays.sort(users, (u1, u2) -> {
-            int l1 = u1.getHighestRole().map(SimpleRole::getLevel).orElse(-1);
-            int l2 = u2.getHighestRole().map(SimpleRole::getLevel).orElse(-1);
-
-            return Integer.compare(l2, l1);
-        });
-
-        for (ChatUser user : users) {
-            Table card = new Table();
-
-            // Avatar
-            if (user.getImageUrl() != null && !user.getImageUrl().isEmpty()) {
-                card.add(new NetworkImage(user.getImageUrl())).size(40 * scale).padRight(8 * scale);
-            } else {
-                card.add(new Image(Icon.players)).size(40 * scale).padRight(8 * scale);
-            }
-
-            // Info Table
-            card.table(info -> {
-                info.left();
-                Label l = info.add(user.getName() + "[white]").minWidth(0).ellipsis(true).style(Styles.defaultLabel)
-                        .color(Color.white)
-                        .left().get();
-                l.setFontScale(scale);
-                info.row();
-
-                user.getHighestRole().ifPresent(role -> {
-                    Label l2 = info.add(role.getId()).minWidth(0).ellipsis(true).style(Styles.defaultLabel)
-                            .color(Color.valueOf(role.getColor()))
-                            .labelAlign(Align.left)
-                            .left()
-                            .get();
-                    l2.setFontScale(scale);
-                });
-            }).growX().left();
-
-            userListTable.add(card).growX().minWidth(0).padBottom(8 * scale).padLeft(8 * scale).padRight(8 * scale)
-                    .row();
-            userListTable.pack();
-        }
-    }
-
-    private void collapse() {
-        ChatConfig.collapsed(true);
-        unreadCount = 0;
-        setup();
-    }
-
-    private void sendMessage() {
-        String content = inputField.getText();
-
-        if (content == null || content.trim().isEmpty()) {
-            Vars.ui.showInfoFade("@chat.empty-content");
-            return;
-        }
-
-        if (currentChannelId == null)
-            return;
-
-        send(() -> ChatService.getInstance().sendMessage(currentChannelId, content, replyingToMessageId)
-                .thenAccept(this::addMessages));
-    }
-
-    private void sendSchematic() {
-        if (currentChannelId == null)
-            return;
-        send(() -> ChatService.getInstance().sendSchematic(currentChannelId, inputField.getText(), replyingToMessageId)
-                .thenAccept(this::addMessages));
-    }
-
-    private void handleAttachContent(String content) {
-        if (content == null || content.isEmpty())
-            return;
-
-        if (currentChannelId == null)
-            return;
-
-        boolean isSchematic = isSchematic(content.trim());
-
-        if (isSchematic) {
-            send(() -> ChatService.getInstance().sendSchematic(currentChannelId, content, replyingToMessageId)
-                    .thenAccept(this::addMessages),
-                    false);
-        } else {
-            send(() -> ChatService.getInstance().sendMessage(currentChannelId, content, replyingToMessageId)
-                    .thenAccept(this::addMessages),
-                    false);
-        }
-    }
-
-    private <T> void send(Prov<CompletableFuture<T>> prov) {
-        send(prov, true);
-    }
-
-    private <T> void send(Prov<CompletableFuture<T>> prov, boolean clearInput) {
-        if (!AuthService.getInstance().isLoggedIn()) {
-            Vars.ui.showInfoFade("@chat.not-logged-in");
-            return;
-        }
-
-        if (isSending.get()) {
-            Vars.ui.showInfoFade("@chat.sending-in-progress");
-            return;
-        }
-
-        try {
-            isSending.set(true);
-
-            if (sendButton != null) {
-                Core.app.post(() -> sendButton.setText("@sending"));
-            }
-
-            prov.get().thenRun(() -> {
-                Core.app.post(() -> {
-                    isSending.set(false);
-                    if (clearInput) {
-                        inputField.setText("");
-                        replyingToMessageId = null;
-                        buildInputTable(inputTable);
-                    }
-
-                    if (sendButton != null) {
-                        sendButton.setText("@chat.send");
-                        if (clearInput) {
-                            inputField.fireClick();
-                        }
-                    }
-                });
-            }).exceptionally((err) -> {
-                Core.app.post(() -> {
-                    isSending.set(false);
-
-                    String errStr = err.toString();
-
-                    if (errStr.contains("409") || err.getMessage().contains("409")) {
-                        Vars.ui.showInfoToast("@chat.rate-limited", 3f);
-                    } else {
-                        Vars.ui.showInfoToast("@chat.send-failed", 3f);
-                        Log.err("Send message failed", err);
-                    }
-
-                    if (sendButton != null) {
-                        sendButton.setText("@chat.send");
-                    }
-
-                });
-                return null;
-            });
-
-        } catch (Exception _e) {
-            isSending.set(false);
-        }
-    }
-
-    private void updateBadge() {
-        if (badgeTable == null)
-            return;
-
-        badgeTable.clear();
-        badgeTable.visible = unreadCount > 0;
-        badgeTable.top().right();
-
-        if (unreadCount > 0) {
-            Table badge = new Table();
-            badge.background(Tex.whiteui);
-            badge.setColor(Color.red);
-
-            Label label = new Label(unreadCount > 99 ? "99+" : String.valueOf(unreadCount));
-            label.setColor(Color.white);
-            label.setFontScale(0.6f);
-
-            badge.add(label).padLeft(2).padRight(2);
-
-            badgeTable.add(badge).height(16).minWidth(16);
-        }
-    }
-
-    private void renderPlayerConnectRoom(Table table, String link) {
-        table.table(t -> {
-            t.left();
-            t.add("@chat.loading-room-info").color(Color.gray);
-
-            playerConnectService.getRoomWithCache(link).thenAccept(room -> {
-                Core.app.post(() -> {
-                    t.clear();
-
-                    if (room != null) {
-                        PlayerConnectRenderer.render(t, room).grow();
-                    } else {
-                        t.add("@chat.room-not-found").color(Color.gray);
-                    }
-                });
-            });
-        }).growX().left().padTop(10);
-    }
-
-    private void renderSchematic(Table table, Schematic schematic) {
-        Button[] sel = { null };
-        sel[0] = table.button(b -> {
-            b.top();
-            b.margin(0f);
-            b.table(buttons -> {
-                buttons.left();
-                buttons.defaults().size(50f);
-
-                ImageButtonStyle style = Styles.emptyi;
-
-                buttons.button(Icon.info, style, () -> Vars.ui.schematics.showInfo(schematic)).tooltip("@info.title")
-                        .growX();
-                buttons.button(Icon.upload, style, () -> Vars.ui.schematics
-                        .showExport(schematic)).tooltip("@editor.export")
-                        .growX();
-                buttons.button(Icon.pencil, style, () -> Vars.ui.schematics
-                        .showEdit(schematic)).tooltip("@schematic.edit").growX();
-
-            }).growX().height(50f);
-            b.row();
-            b.stack(new SchematicImage(schematic).setScaling(Scaling.fit), new Table(n -> {
-                n.top();
-                n.table(Styles.black3, c -> {
-                    Label label = c.add(schematic.name()).style(Styles.outlineLabel).color(Color.white).top().growX()
-                            .maxWidth(200f - 8f).get();
-                    label.setEllipsis(true);
-                    label.setAlignment(Align.center);
-                }).growX().margin(1).pad(4).maxWidth(Scl.scl(200f - 8f)).padBottom(0);
-            })).size(200f);
-        }, () -> {
-            if (sel[0].childrenPressed())
-                return;
-            if (Vars.state.isMenu()) {
-                Vars.ui.schematics.showInfo(schematic);
-            } else {
-                if (!Vars.state.rules.schematicsAllowed) {
-                    Vars.ui.showInfo("@schematic.disabled");
-                } else {
-                    Vars.control.input.useSchematic(schematic);
-                }
-            }
-        }).top().left().pad(4).style(Styles.flati).get();
-
-        sel[0].getStyle().up = Tex.pane;
-    }
-}
